[ Feat ] subsidiary component on sign-in
